"8:30am. I think I accidentally refed the F# issue in yesterday's reply. Damn. That will create a link in it.
It was late so I wasn't cautious about it. This is one of the disadvantages of using Github as your blogging platform.

Anyway, I've chilled so let me get started for the day.

10:15am. Focus me stop reading the stuff on the ML sub.

///

Sigh, what we will do is put this project aside for the moment.
Let's start a new project that will be based off the MudBlazor template, and then we will try importing the diagrams component.
We want to make sure there isn't some fundamental incompat

///

Damn it, I do not want to start a new thing.

Let me pause here and I'll get to the bottom of it.

1pm. Done breakfast.

Bolero has kicked my ass thoroughly. I am going to switch to Blazor after all.

I can't even create a regular button in its DSL.

There is no point to using Bolero if I can't get Blazor components to work with it.

1:45pm. Let me resume.

https://fsbolero.io/docs/HTML

```
let myElement : Node =
    button {
        on.click (fun _ -> printfn "Clicked!")

        "Click me!"
    }
```

Ah, here is how it is possible to do it.

Enjoy Yourself by Steven O'Brien | https://www.steven-obrien.net/
Creative Commons CC BY-ND 4.0
Music promoted by https://www.chosic.com/free-music/all/
https://creativecommons.org/licenses/by-nd/4.0/

https://soundcloud.com/yoitrax/latte-by-jantrax-royalty-free-music-original
https://soundcloud.com/liqwyd/taste-it

I should leave links for these in the media clips.

5:25pm. https://youtu.be/m8Te-3rTopI
VN Compiler. We give up on Bolero. Introducing MudBlazor. (Pt. 2)

Let me post this on the F# sub.

#webdevelopment #fsharp #blazor #AspNetCore

Let me save these tags here.

5:35pm. I am done for the day here.

5:55pm. I guess I'll read more abyss loli manga.

https://mangadex.org/title/bf1d564b-e41a-4f55-9241-b5c7ac7f30ab/despite-coming-from-the-abyss-i-will-save-humanity

Let me link to it again. It really does pick up after the village battle surpringly enough and is a fun read. The novel is supposed to be GB, but the manga makes zero mention of it.

Sigh, forget it. I'll just focus on this project until I feel like I have perfected my webdev workflow which shouldn't take more than a few months once I start focusing on Blazor.

I meant to go with React, but it is not my job to specialize in outdated technologies.

I'll learn the best tools, and if the companies aren't interested that is their problem.

I'll probably end up liking Blazor.

I'll host the Leduc project properly, and put it on my resume along with Helix and Spiral, and hopefully that should be enough to get me an in.

I've already put the Youtube series on my resume. Out of the 20 compos I've applied to, I got 3 rejects so far, and I don't get the feeling I'll get any interest from the rest.

Most likely I will get interest once I get enough of an audience on Youtube, and then I'll have proper leverage for negotiation.

But that is going to be a slow process.

Still, I've realized it, had I done writing, even at the placid pace I've been getting followers, I'd have been way more successful than had I been trading or programming.

It is important to keep putting new material out there, and cultivate your connections.

Right now I have 106 subscribers on Youtube, which is not a big deal.

But it is better than Heaven's Key, and gives me a platform to cultivate my skills.

Getting a little bit like this every day is far better than programming in silence and solitude even if it slows me down.

It is not like I'll get anything in particular from the Helix project.

I've gone from living in a world where I life with my own power to, living in a world where I am dependent on the attention of others.

Because the vision of the Singularity that I had did not manifest.

The world just can't throw me a line here, but I still want to believe that programming will have a purpose in the end.

Street cleaners are not going to be the ones to bring the era of a man to an end.

I still want to believe that the programmers and the individuals will be the ones to do it.

I still want to believe that I can do something for myself with my own power.

I do not want to give up."