"7:40am. Let me just take my turn in the bathroom and then I will start.
https://youtu.be/Q1zwnRB-B2c
BEST EDM SONFS BY LiQWYD

I really dig LiQWYD's music.

8:05am. Let me chill a while longer. I guess the Blazor arc will last a while until I get used to it. Now that I've gone into Blazor my desire to learn JS frameworks has dropped to bottom.

I'll focus on what I am good at.

8:15am. Damn it, that comment I wrote yesterday as a reply did not get posted at all. WTF?

9:40am. Let me start. Done with the bath.

My comment got eaten again.

> Sigh, I wrote a long reply and it got eaten twice already. In short, It is pretty good. Check out out my channel to see me using it. Ask me again there, and I'll answer you properly without the spam filter getting in our way.

9:50am. I'll check out later whether this comment stuck.

Let's start work for the day.

https://soundcloud.com/liqwyd/sweet
https://soundcloud.com/liqwyd/night-out-free-download
https://soundcloud.com/liqwyd/mellow
https://soundcloud.com/liqwyd/sunrise
https://soundcloud.com/liqwyd/deeper

11:15pm. My god, my replies to that guy are just getting eaten.

12:30pm. Finally done with the video. Let me write the outro.

///

Let's take a break here to review. We've been programming since uncharacteristically early today and are tired by now.
We finally managed to bring back F#, so now this video deserves its place on the F# playlist.
We implemented that last condition that the inputs can have only one link, but we think we are going to get rid of it.
Visual novels can have branches so it might be good to have merging behavior in text nodes.
It would also make it easier to demonstrate why we need memoization for the graph traversal.
So we'll go with that decision, and comment out the condition, but we'll keep the code around for later.
We'll likely need it.
I'll let you in on a little secret, even though this is a VN compiler for now, master has his sights set high.
He is pretty envious of the comfy anon, so who knows what kinds of crazy nodes we might have in the future.
He picked a simple goal that can do something useful and inline with his own interests to start him off, but this project has the potential for long term cultivation into a visual programming language in its own right.
We can only hope he won't be wrong again, and we won't have the next 10 years dominated by Stable Diffusion.
With web development we should be going in the right direction, right?

///

I'll never get the chance to post this from yesterday. Nevermind it.

12:55pm. 25m. Let me go get breakfast while it renders.

2:20pm. https://mangadex.org/title/bf1d564b-e41a-4f55-9241-b5c7ac7f30ab/despite-coming-from-the-abyss-i-will-save-humanity

Let me read the latest chapter of this and I will resume.

2:35pm. https://docs.scala-lang.org/scala3/new-in-scala3.html

I am reading up Scala. It has a lot of advanced capabilities.

I guess I'd pick it if I was a JVM programmer.

Anyway, let me resume. I need to make the thumbnail for the video and publish it.

https://youtu.be/crqGY9__k3o
VN Compiler. Creating The Compilation Node. Compiling The Textual Nodes. (Pt. 3.2)

Whatever, here it is.

#fsharp #blazor #aspnetcore #dotnet #csharp #webdevelopment #nodeui

3:10pm. https://www.youtube.com/playlist?list=PL04PGV4cTuIUDwcf_JwUkO5rBAo6LFQ9W
Blazor And F# For VN Compilation

Let me post this playlist on the PL sub.

https://www.reddit.com/r/ProgrammingLanguages/comments/13scv3o/blazor_and_f_for_vn_compilation/

Maybe it will turn out to be 0 votes much like with the time I posted a vid to the Typescript sub, but whatever.

3:15pm. https://mangadex.org/chapter/3096dffa-3d7e-4208-9a9d-7bff86f9997c/1

Let me finally read this, and then I will get started on the next thing.

4:05pm. Let me resume for real.

4:40pm. https://learn.microsoft.com/en-us/training/modules/publish-app-service-static-web-app-api-dotnet/

As expected it supports Blazor WebAssembly.

4:45pm. https://learn.microsoft.com/en-us/training/modules/publish-app-service-static-web-app-api-dotnet/

I'll go over these later.

Sigh, this isn't really a time to be starting a new project. Since it is almost 5pm, I should be ready to stop.

And I am starting to feel fatigue.

https://mudblazor.com/components/linechart#basic-usage

I guess I'll be bringing in MudBlazor for this project as well.

///

The blazor.webassembly.js script is provided by the framework and handles:

* Downloading the .NET runtime, the app, and the app's dependencies.
* Initialization of the runtime to run the app.

///

I'll just read up a bit on this and then call it a day. I definitely do not feel like diving into it.

> There's no .NET server-side dependency after the app is downloaded from the server, so the app remains functional if the server goes offline.

I didn't know that.

5:25pm. Let me close here. I'll go have lunch.

I am just wondering where the .NET runtime would be hosted when I deploy my app. But I'd guess at my own site. Probably not at MS's"