"https://news.ycombinator.com/item?id=36100808
///

Hello,
I'm an HN user for 5-6 years but I wanted to use an alt for this as it's a bit sensitive.

Last week I created a ChatGPT clone for my city. It's basically a laravel app hooked up to GPT3.5-turbo with a custom prompt that makes local references and talks in the way that people in my city talk.

It's been a surprise hit and now I've had 20,000 chats coming through in one day - people where I live really seem to love it

The problem is that I likely can't afford to keep hosting this. It's cost me $50/day for one day, and Adsense doesn't allow 'chat apps', so I'm at a loss at how to cover the bill for this app. I've already optimised the prompt and reduced the number of tokens I'm sending

The app a joke really, but it's a local joke that seems to be quite popular and connecting with my city. Should I try to raise donations? Is there an advertising provider I could use that would potentially cover the costs? Is there an alternative to OpenAI that is comparable to GPT-3.5 that I could self-host cheaper?

Any and all advice appreciated. I don't care about profit - I just want to keep the app online so people can enjoy it.

///

There are hints all around how to make money. But damn, do I feel like not doing anything in this direction. I have zero money, literally zero.

If I had a thousand bucks to absorb in initial hit these kinds of projects would be doable.

https://soundcloud.com/anttimartikainen/albums

Damn, this guy is good.

> hackernews has monthly threads where candidates showcase themselves for employers, but I forgot what it was called. Do you know?

https://hnhiring.com/technologies/csharp

https://news.ycombinator.com/item?id=32619199

> I left my first sysadmin job after about 3 years because the advice I recieved is that if you work solo for too long, you develop your own ways of doing things, become unable to integrate into more mainstream best practices, and become effectively "unhirable". I wonder if that has something to do with your lack of success in the job hunt.

The world is rough.

5/29/2023

7:40am. Let me chill for a while and then I will start.

8:25am. Let me start.

https://news.ycombinator.com/item?id=24758772

///

Thank you, serverless is the new structure that everyone loves and wants to work in, but no one talks about the downside. I just took over a project that has a serverless architecture.
We can't run it locally because the emulator won't run 7/10 of the code, there are memory limit errors that are totally opaque to us since we can't step through and watch it break. We are getting unexplained timing issues that were being worked on via logs in dev. The costs are insane.

If we want to say serverless is the future, the future isn't here yet. There are many tools that would need to be built to make serverless viable, and the tools that are built are immature and bad. Not to mention, you're tied to the companies support, so Google was outdated from the LTS for Node for a while. The reality is going serverless means you don't have control over what happens to your code.

///

///

I believe one of the main reasons Serverless will never be more than a low-tier niche is the combination of the following:
In the end, you are just renting well maintained server farms (well, a specific percentage of operational time of some of the servers in them). There is absolutely no appeal for large technology-based companies do this once they can (the following is the lowest scale example) afford to maintain their own servers, while potentially renting a few offsite backups in other areas of the world (again, this is just the lowest-scale architecture starting at which Serverless is always the worse option).

Existing solutions are extremely over-engineered. It can be excused with "officially planning for the use-case where maximum scalability is required", but it's almost certainly a pretense to sell "certifications", aka "explaining our own convoluted badly documented mess". What this actually means is that many SWE's who are good enough to learn to use Serverless effectively, can learn any other framework that allows building distributed systems across server nodes with equal effort. Why would I base my whole business on your vendor locked, sub-optimized dumpster when I can do the same on an infinitely scalable VM networks, that can be ported to literally any vendor who supports $5/mo VMs (or, you know, self hosted if my company is large enough)?

///

Enough of this crap. Let me start for real.

10am. https://soundcloud.com/anttimartikainen/albums

All of this guy's music is so good! Right now I am listening to Northern Steel.

10:40am. Had to take a break. I figured out how I want to do this.

12:15pm. Let me take a break here.

1:10pm. Let me resume. I want to get programming part at least done today.

```
E:\Webdev\Fable\CFR-In-Fable\src\Lproj.UI\Microsoft.NET.Sdk.Razor.SourceGenerators\Microsoft.NET.Sdk.Razor.SourceGenerators.RazorSourceGenerator\Components_TrainUI_razor.g.cs(213,51): error CS0030: Cannot convert type 'System.F
unc<System.Tuple<int, System.Tuple<double, double>>, decimal>' to 'System.Func<System.Tuple<int, System.Tuple<double, double>>, decimal?>' [E:\Webdev\Fable\CFR-In-Fable\src\Lproj.UI\Lproj.UI.csproj]
```

What is this horseshit?

4:40pm. I am stuck, I stumbled upon a really vicious concurrency bug, and the mailbox processor isn't working as expected at all.

4:45pm.

```
            let mb =
                new MailboxProcessor<_>(fun mb -> async {
                    let mutable model = PlayServerModel.Init
                    while true do
                        let! msg = mb.Receive()
                        model <- update msg model mb.Post
                    })
            mb.Start()
            mb
```

I thought of removing the cancellation token and it started working!

```
            printfn "is canc: %b" cancellation_token.IsCancellationRequested
            cancellation_token.ThrowIfCancellationRequested(
```

This is weird. It is not like it gets cancelled.

Ah whatever. Let me make a minimal repro and I'll open an issue.

https://github.com/mrakgr/CFR-In-Fable/tree/mailbox_bug_dotnet7
https://github.com/mrakgr/CFR-In-Fable/tree/mailbox_bug_dotnet8_preview3

```
    do srv.Post("Hello from view")
    do this.Say("123456789")

    member _.Say(x) = srv.Post x
```

Wait, I just realized. This second line does not work.

5:30pm. Open an issue on the F# repo.

5:55pm.

https://www.youtube.com/watch?v=iLkwUJBsf6g
Try now! Experimental multi-threading support for Blazor WebAssembly (coming in .NET 8)

This only arrived in .NET 8.

https://youtu.be/iLkwUJBsf6g?t=35

Oh, I have to enable WASM threads.

https://visualstudiomagazine.com/articles/2022/10/11/blazor-webassembly-net7.aspx

https://youtu.be/iLkwUJBsf6g?t=346

Mhhh, do I need this JS code?

https://youtu.be/iLkwUJBsf6g?t=424

Here he is talking about needing to enable some things.

6:30pm. https://github.com/dotnet/runtime/issues/

No, it is still being worked on. Issue 68162, and others.

I have no choice, I'll have to switch to Blazor Server.

There simply isn't much I can do with this right now.

Let me go get lunch."