"8:15pm. I am so braindead right now. Edge keeps talking about Split Screen, but I can't find where that is.
8:45pm. https://boards.4channel.org/a/thread/253286943/why-did-scifi-anime-lose-its-imaginative-flair#p253333948

///

I'd like to add this too - isekai is the actual dream and the actual longing of current Humanity as a whole. A return to a freer, more natural existence. Rural or medieval settings, no surveillance tech, nothing that distracts or obstructs life between people in all its aspects. The simple fact is that progress and technology did not liberate Mankind, did not give us more freedom or anything of the sort. Everything has been taken from us - even our imagination (ChatGPT, AI as a whole). The omnipresent tyranical state, be it the US, EU or China, it doesn't matter, and neither does the ideology since the end result is the same - total surveillance, social credit (openly or subtly implemented), censorship, and finally slavery. What underlies the death of SciFi is the observable fact that technological progress has only made us into slaves of the state.

Space travel? Moonbases? Interstellar travel? None of those things actually matter or will make our lives better, but worse. Technology has failed, utterly failed to help Humanity in a fundamantal sense, it only made everything worse as can be seen from everything from incels to birthrates to crime to mental health issues. We ARE living the SciFi dreams of yesterday, and its a nightmare. The Japanese probably feel this even more acutely, ergo isekai in fictionalized medieval Europe. The only reason Chinese SciFi exists is probably because it serves the goals of the State, and that's about it. It's their equivalent of woke propaganda and the rainbow flag.

Anyone who still advocates for technological progress or dumb shit like space travel and so on is an idiot in my eyes. The future, a real future, is distinctly rural and with PC's running the equivalent of Win98 with a dial up connection, but you spend most of your time fishing in the local lake and hanging out with your fellow peasant friends while you drink and discuss your fields and crops. THAT is a future worth fighting for and striving for.

///

I want the opposite of this. I want the Singularity!

https://mangadex.org/title/3eaabe03-d2bd-43f3-af71-bd1db4ab564a/serial-killer-isekai-ni-oritatsu

What is with this manga and all these shit tier translators picking it up on a whim?

6/5/2023

7:35am. Damn it, I couldn't get enough rest. Overwork takes its toll. But I just couldn't leave the issue alone. Let me set the render for yesterday's vid.

7:40am. I just didn't have the energy to finish it yesterday.

9am. https://youtu.be/ExE6YDR_MtE
VN Compiler. Implement the undo/redo functionality. (Pt. 3.5)

I'll come back in a few hours and publish it. Actually let me schedule it.

https://mangadex.org/title/59e04a79-435e-418d-8b64-c8d4315693ab/akuma-koujo

I'll leave this for later.

https://www.youtube.com/watch?v=67TkpHKv9tE
YouTube Algorithm 2023: What Creators Need to Know

Also, I still haven't watched this. I'll leave it for later. I want to start programming.

9:05am. Let me start the programming session for the day.

11:40am. I am getting distracted, this is a good spot to stop and have breakfast.

I haven't really excelled. Maybe it is a good thing to do a regular screencasting simply for the sake of momentum.

Now that I've showcased how to deal with undo-es, it is time to get to the point.

I need to save and load the images to blob storage. Also to disk.

11:50am. Let me stop here."