I am trying to do a wholesale upgrade of the project to SignalR and .NET 7 and it is kicking my ass hard.
At the very least, I at least managed to figure out how to connect from one end to the other and now messages are getting through.

I am actually trying to do a screencast of this, and it is turning into a huge mess, no way I can release this.

Eh, I suppose things like these are expected when tackling complex things. Software dev is not roses all the time.

I'll dive into it and figure out why only empty messages are being sent back tomorrow, it is too late in the day now. I hope it is not the case that Hopac is broken on .NET 7.

Note: Ammend exists only to mess up your repo.