"https://arstechnica.com/gaming/2019/04/coding-without-a-keystroke-the-hands-free-creation-of-a-full-video-game/
https://xeiaso.net/blog/voice-control-talon/

sleep mode - in this mode, Talon is not listening for commands
command mode - in this mode, Talon listens for commands and executes them in the order they are spoken
dictation mode - in this mode, Talon listens for large blocks of text and types them out as you say them

It has a dictation mode? I'll have to look into that.

11/13/2023

9:05 AM.

///

2023-11-13 13:24
HK, International flight has departed

2023-11-12 19:32
CN, Arrived at the origin international airport

2023-11-12 09:32
Mainland China, CN, Shipment is in transit to next facility

2023-11-12 00:19
Mainland China, CN, Departed from sort facility

2023-11-10 20:50
Mainland China, CN, Arrived at origin facility

2023-11-10 16:07
Mainland China, CN, Shipment picked up

2023-11-10 14:02
Sesvete, HR, We are exchanging data internally

2023-11-07 15:53
Shipment information received

///

My glove 80 keyboard finally departed from the Hong Kong Airport.

I don't know what the target destination is, so I can only hope it's Croatian. If it's that, then I'll definitely get the keyboard this week.

Today, my goal will be to go through the practice exercises for Talon.

After I feel secure at that, I'm going to do a real programming session. The reciprocal isn't much to it. I have to get used to using my voice to program. And the best way to do that is to temper myself with a real assignment.

Using voice typing to do this journal is really a good exercise for talking.

I wish I knew about it earlier because I definitely would have used that had I knew how good it was.

My ring finger is a covering because it's definitely not as number as it was in the past few weeks. Even so, the recovery will take a few months.

The keyboard once it gets here won't make a difference in terms of allowing me to return to my previous fitness. Rather, it will allow me to use my left hand better. I will be able to use it for a navigation with a new keyboard.

12:05 PM. https://kbd.news/Glove80-Rethinking-split-contoured-ergonomic-keyboard-1796.html?utm_source=Klaviyo&utm_medium=flow&_kx=y-P3vN0awMgfN_5-fSYV6hn3CWnqOweQI9jK0_2a9oI%3D.V4FbQN

I got an email from MoErgo, so now I'm reading this article.

12:55 PM. Let me do the chores, and then I'll figure out how Talon configuration files work. Before I do anything with it, I need to figure out how macros can be used using Talon. I could try to make do without it, but it would be a big disadvantage to do it that way.

After that I might do some like programming, but not an actual session. I need, I need to give my right hand time to rest. If I start a program for real now, it will put a stress on it.

1:30 PM. His time to resume. I'll try making my own macro using Talon configuration files. For my first move, I'll check out the community user config files.

inl _ =:3
inl _ =:3

Ohh wow. The macro actually worked, though not correctly...

inl _ =:3

inl _ =

inl _ =
inl some_variable =
inl someVariable =

Let me pace the macro here so you can see how it is made.

///

-

inline:
    insert("inl _ = ")
    key(left:3)
    key(ctrl-d)

///

I like this. This will be useful definitely.

I just got to put the Talon file into the directory, and I didn't even need to restart the application in order for the macro to take effect.

It feels like I've met one of the goals for today.

What I should do right now is study a bit more about how the macros really work. Let me do that right now.

After I clear that goal, i'll be able to start programming using talon.

`date insert: insert(user.time_format("%Y-%m-%d"))`

I just found this in the dates Talon file. I'm going to edit it so it inserts the date using my usual format.

13/11/2023

`date insert: insert(user.time_format("%d/%m/%Y"))`

It works perfectly now.

13/11/2023

THIS WORKS!!!!

2:30 PM. No, time for a break. After that, after that, I'll do some more research on how our arguments work in talon files.

https://www.joshwcomeau.com/blog/hands-free-coding/

> Formatters can be composed. For example, I'm a fan of UPPER_SNAKE for JavaScript constants:

I didn't know they could be composed.

3:50 PM. I can't get the upper snake to work. But I did manage to compose some of the others. It's no big deal.

There are a lot of interesting commands in the community folder. Right now I am taking a look through it.

In particular, there is a lot of stuff for vs code.

4:05 PM. I am done surveying the commands. The next thing I'm going to do is go through pokies Youtube channel.

After that I'm going to be ready for what will come ahead.

5:45 PM. His channel is pretty boring actually. So I'm going to skip Cursorless. Cursorless has advanced functionality and the Talon can barely understand me as it is. So I'm going to navigate using the keyboard and the mouse. Talon will be there just so I can offload some of the input burden. I first need to master Talon before trying to go beyond it.

And at any rate, the wise thing to do wouldn't be to try to get back into programming right away.

Right now I'm in torn whether I want to go back to writing Simulacrum or whether I want to continue programming. I really should try to just maybe get the first chapter out. Just focus on that and I'll see how I feel about getting back to programming after that. Everything is **** **** right now, so I can at least spare this much.

Let's just give it a day or two or three just to get the first edge a chapter out or even just the start of the first chapter. If I really don't feel like it, then fine. I'll go back to programming, but if I do feel like writing, then I should write until the hand recovers. It's not like I am a crunch for time. The job market is still messed up, so it's not like I'm missing much by not applying to jobs or doing my own projects.

I pretty much wasted the last 5 and 6 weeks, but this is the one situation where not programming is actually justified. This is pretty much the perfect time to take a vacation, a prolonged one.

I can regret not noticing the warning signs of rsi earlier. They were there with me for years. And that's actually the reason why I ignored them. I've been able to ignore them for years, so I thought that I would be able to ignore them for a few months more. If I could go back, I'd tell the old me to maybe buy a trackball mouse and use voice dictation for writing the journal and as well as writing those screencasts. That would have cut the amount of text I've been writing by 80%. And if that was the case, I probably wouldn't have gotten RSI.

But going back more than three months ago I wasn't exactly swimming in money. It was only recently that I had a little bit of extra cash lying around to upgrade my rig, as well as buy the ergonomic setup.

If this is the end of my path, then so be it.

In any other era, I would have just spent my time writing fiction dreaming about the Singularity."