"9:15 AM. No update for my keyboard. It seems it is still in Hungary.
https://youtu.be/YWimz6rfuls
This is 99% why your aim is bad.

https://youtu.be/gNDfH7J2DNE
Fix Your Terrible Mouse Grip NATURALLY

Today, what I'm doing is trying out the claw grip, even on the track ball mouse.

I'm going to try letting my fist stay in the shape of a clove. In other words, it's a fully relaxed posture. I want to see whether that makes any difference to my rsi. One thing I noticed since I am actually using my whole arm tool move the mouse cursor. It's a lot more accurate than just using my thumb.

https://youtu.be/5S9iwYnndp8
The Optimal Contact Points For Aiming

Right now entering out the claw grip. And it's really **** amazing how much more accurate I am. Instead of using the damp, I'm rotating my wrist to move the mouse cursor. The reason why I couldn't do that before and why I've been using the pole grip for 20 years is because of that a desk. It has a very, very.very little space between the keyboard and the mouse and the top of the ceiling. By ceiling, I mean, it's kind of like a pull out desk, which has a board in the middle. Now that I'm using a regular desk, I have a lot more space above for my hand to move. I should try typing that way as well.

I have a hypothesis that my whole rsi in the right hand isn't so much because of typing, though the volume I put in certainly contributed to it. But because I've been using the palm grip on both the mouse and the keyboard.

https://youtu.be/5S9iwYnndp8?t=50
> When we grip the mouse with the tips, we gain dexterity.

I've been using the mouse for 20 years, and this is the first time discovering this.

https://youtu.be/5S9iwYnndp8?t=145
> If someone you know has a messy handwriting, have a look at their grip. You'll be surprised to know how much of one of their fingers or more are hyper extended.

https://youtu.be/qh7fqjNqCLw
Fixing X-Pro Gamer's Worst Habit (You Do This Too)

Now that I've changed my grip It feels like I am flowing with this track ball. It is so much better now.

10:45 AM. I think I figured it out. All the flaws in my technique. The reason I have an size is probably the way I hold my hands. My left hand is mostly in deflection position by all my right hand is extended. That is how I use the keyboard in general. And that is how I use the mouse.

I think once I adjust my grip, the other side should go away. So I'll need to test it for another few weeks.

Using voice dictation for most of my unstructured text input as well as an ergonomic mouse and the keyboard I'm going to get should be enough to fix it.

https://deepmind.google/discover/blog/transforming-the-future-of-music-creation/

Music generation. Now this is something I've been waiting for a long time.

https://blog.jim-nielsen.com/2023/html-web-components-an-example/

https://youtu.be/R4Ri4ft7bXY
The Good, The Bad, and The Web Components - Zach Leatherman | JSHeroes 2023

3:00 PM. https://youtu.be/C8RtH2eoBU8?list=PL8JmJYDyXnXlrwJPh3ve11lxhNBXQGw64
Witch Weapon Story [CHAPTER 8] [PART 2]

3:10 PM. Let me check on just a bit and then I'll start writing. I've been resting too much.

3:30 PM. Right now I'm not using the keyboard with my right hand too much. I mostly focusing on the left. Thinking about it objectively, I'm doing a lot of weird movements and contortions, even made by a left hand. My typing technique is pretty messed up either way. I think that it's not worth fixing. The keyboard, the regular keyboard that I have is pretty badly designed So it's just so the way I'm typing is just a consequence of that.

Once the glove ate the keyboard arrives, I'll be able to correct my typing.

And right now...

///

Package status: In transit (8 Days)
Country: China -> Croatia
2023-11-15 18:00 HU, International flight has arrived
2023-11-13 13:24 HK, International flight has departed
2023-11-12 19:32 CN, Arrived at the origin international airport
2023-11-12 09:32 Mainland China, CN, Shipment is in transit to next facility
2023-11-12 00:19 Mainland China, CN, Departed from sort facility
2023-11-10 20:50 Mainland China, CN, Arrived at origin facility
2023-11-10 16:07 Mainland China, CN, Shipment picked up
2023-11-10 14:02 Sesvete, HR, We are exchanging data internally
2023-11-07 15:53 Shipment information received
======================================
Powered by www.17track.net

///

It seems that the package is still in Hungary. It's really thinking about to leave from the airport. I thought that would have been done long ago. So far I've been waiting for it to 10 days. Once I get the keyboard, I'll do a programming session. So far my rsi has already taken six weeks from me. Maybe the keyboard will create a miracle. If not, I'll have to learn how to program using a combination of typing and voice programming.

7:25 PM. Word Count's at 8.9 K.

Unlikely the previous chapters I'm going to step up posting these as zip files. It's a waste of time. Instead, I'll just share a link to the Google Docs document.

I'll prove read it tomorrow and then post it. Tomorrow is Saturday, so I have work."