That's a programming today. It seems none of the host functions. And that includes even the host array can be used in a device code. I thought of doing this part in a screencast, but it's too difficult. So far, voice programming is more like a band aid than an actual solution. Things that used to be easy for me are now difficult. And it feels like doing voice programming is playing around.
Because the voice software gets the words wrong. Always, I cannot really rely on it. And it really slows my pace down.

I guess I'm going to have to get some exercise. Training wire upper arms is the next thing on my agenda.

At any rate, it seems that a merge sort that I did for the fpga in the previous sessions will be seeing some use.

The cue that stand on the library barrel has anything in it. Cuda and bound does have some sorting functions, but they are blockwise and work wise, so they aren't what I need here.

I've done a lot of thinking about what I want to do next and I'm going to drop the screencast and I'm going to drop some of the reinforcement learning methods and the evolutionary methods. Instead, I'm going to be honest to my desires and go straight to what I want to do, which is genetic programming. I have a job so I don't feel like messing with making poker boats. Instead, I'm going to use this whole thing to try and conquer the absence of reason.